expand example app to demonstrate attatch to mesh
